requirements.txt
torch
torchvision
h5py
pydicom
six
numpy
scipy
Pillow
matplotlib
scikit-image
opencv-python
imageio
Shapely
imgaug